# evidencias/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "device_systems API",
    "description": "API REST para la gestión de usuarios de device_systems. Utiliza FastAPI, Pydantic v2 y SQLAlchemy para persistencia relacional, CRUD completo, filtros, validaciones, constraints y manejo controlado de errores. Migraciones Alembic, dispositivos y préstamos.",
    "contact": {
      "name": "Nombre Apellido",
      "email": "ejemplo@device-systems.com"
    },
    "version": "4.0.0"
  },
  "paths": {
...
    "/users/{user_id}/devices": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Dispositivos asignados",
        "description": "Equipos con préstamos active u overdue del usuario.",
        "operationId": "user_devices_users__user_id__devices_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dispositivos actualmente asignados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceResponse"
                  },
                  "title": "Response User Devices Users  User Id  Devices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/devices": {
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "Listar dispositivos",
        "description": "Combina filtros por tipo, disponibilidad, marca y búsqueda por nombre o serie.",
        "operationId": "list_devices_devices_get",
        "parameters": [
          {
            "name": "device_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Type"
            }
          },
          {
            "name": "is_available",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Available"
            }
          },
          {
            "name": "brand",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Brand"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dispositivos que coinciden con los filtros",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceResponse"
                  },
                  "title": "Response List Devices Devices Get"
                }
              }
            }
          },
          "404": {
            "description": "Dispositivo inexistente"
          },
          "400": {
            "description": "Serie duplicada o actualización vacía"
          },
          "409": {
            "description": "Dispositivo con préstamos"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Devices"
        ],
        "summary": "Crear dispositivo",
        "description": "Registra un equipo con número de serie único.",
        "operationId": "create_device_devices_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dispositivo registrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Dispositivo inexistente"
          },
          "400": {
            "description": "Serie duplicada o actualización vacía"
          },
          "409": {
            "description": "Dispositivo con préstamos"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/devices/{device_id}": {
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "Consultar dispositivo",
        "description": "Obtiene un equipo por su identificador.",
        "operationId": "get_device_devices__device_id__get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos del dispositivo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Dispositivo inexistente"
          },
          "400": {
            "description": "Serie duplicada o actualización vacía"
          },
          "409": {
            "description": "Dispositivo con préstamos"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Devices"
        ],
        "summary": "Reemplazar dispositivo",
        "description": "Reemplaza los campos de un equipo sin préstamos abiertos.",
        "operationId": "replace_device_devices__device_id__put",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dispositivo actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Dispositivo inexistente"
          },
          "400": {
            "description": "Serie duplicada o actualización vacía"
          },
          "409": {
            "description": "Dispositivo con préstamos"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Devices"
        ],
        "summary": "Modificar dispositivo",
        "description": "Actualiza solo los campos enviados de un equipo sin préstamos abiertos.",
        "operationId": "patch_device_devices__device_id__patch",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevicePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dispositivo actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Dispositivo inexistente"
          },
          "400": {
            "description": "Serie duplicada o actualización vacía"
          },
          "409": {
            "description": "Dispositivo con préstamos"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Devices"
        ],
        "summary": "Eliminar dispositivo",
        "description": "Elimina un equipo únicamente si no tiene historial de préstamos.",
        "operationId": "delete_device_devices__device_id__delete",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminación sin contenido"
          },
          "404": {
            "description": "Dispositivo inexistente"
          },
          "400": {
            "description": "Serie duplicada o actualización vacía"
          },
          "409": {
            "description": "Dispositivo con préstamos"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/devices/{device_id}/loans": {
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "Historial del dispositivo",
        "description": "Consulta con joins todos los préstamos históricos de un equipo.",
        "operationId": "device_loans_devices__device_id__loans_get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Préstamos con usuario y dispositivo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LoanDetailResponse"
                  },
                  "title": "Response Device Loans Devices  Device Id  Loans Get"
                }
              }
            }
          },
          "404": {
            "description": "Dispositivo inexistente"
          },
          "400": {
            "description": "Serie duplicada o actualización vacía"
          },
          "409": {
            "description": "Dispositivo con préstamos"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DeviceCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "serial_number": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Serial Number"
          },
          "device_type": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Device Type"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "is_available": {
            "type": "boolean",
            "title": "Is Available",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "serial_number",
          "device_type"
        ],
        "title": "DeviceCreate",
        "example": {
          "brand": "lenovo",
          "device_type": "laptop",
          "name": "Lenovo ThinkPad",
          "serial_number": "LEN-001"
        }
      },
      "DevicePatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "serial_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Serial Number"
          },
          "device_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Type"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "is_available": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Available"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DevicePatch"
      },
      "DeviceResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "serial_number": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Serial Number"
          },
          "device_type": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Device Type"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "is_available": {
            "type": "boolean",
            "title": "Is Available",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "serial_number",
          "device_type",
          "id",
          "created_at"
        ],
        "title": "DeviceResponse",
        "example": {
          "brand": "lenovo",
          "device_type": "laptop",
          "name": "Lenovo ThinkPad",
          "serial_number": "LEN-001"
        }
      },
      "DeviceUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "serial_number": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Serial Number"
          },
          "device_type": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Device Type"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "is_available": {
            "type": "boolean",
            "title": "Is Available"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "serial_number",
          "device_type",
          "is_available"
        ],
        "title": "DeviceUpdate",
        "example": {
          "brand": "lenovo",
          "device_type": "laptop",
          "name": "Lenovo ThinkPad",
          "serial_number": "LEN-001"
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoanDetailResponse": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/LoanStatus"
          },
          "return_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Date"
          },
          "user_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "User Id"
          },
          "device_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Device Id"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "loan_date": {
            "type": "string",
            "format": "date-time",
            "title": "Loan Date"
          },
          "user": {
            "$ref": "#/components/schemas/UserPublic"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceResponse"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "return_date",
          "user_id",
          "device_id",
          "id",
          "loan_date",
          "user",
          "device"
        ],
        "title": "LoanDetailResponse",
        "example": {
          "device_id": 1,
          "user_id": 1
        }
      },
...
      "LoanStatus": {
        "type": "string",
        "enum": [
          "active",
          "returned",
          "overdue"
        ],
        "title": "LoanStatus"
      },
...
      "UserPublic": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 3,
            "title": "Name",
            "description": "Nombre completo del usuario. Mínimo 3 caracteres.",
            "examples": [
              "Usuario Ejemplo"
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Correo electrónico único del usuario.",
            "examples": [
              "usuario@ejemplo.com"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "description": "Rol del usuario dentro del sistema: admin, support o user.",
            "default": "user"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Indica si el usuario se encuentra activo en el sistema.",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "id"
        ],
        "title": "UserPublic",
        "description": "Response model utilizado por los endpoints de la API pública.",
        "example": {
          "email": "usuario@ejemplo.com",
          "id": 1,
          "is_active": true,
          "name": "Usuario Ejemplo",
          "role": "admin"
        }
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "admin",
          "support",
          "user"
        ],
        "title": "UserRole",
        "description": "Roles permitidos para un usuario del sistema device_systems."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}